Rename minimalParser to parser
